UPDATED DATASET WITH ENTSOE 2022 DATA
This update contains all the modifications applied during the
first revision of the paper. The source for RES data is ENTSOE
PEMMDB, where ENTSOE provide projections of hydro, onshore,
offshore, PV, CSP and demand progiles for 2025 and 2050.
Generation and transmission capacities in 2030 are provided in
the latest relaease of the 2022 TYNDP.